+++ Mythras/templates/scripts/base.js
@@ -221,8 +221,12 @@
     
     // Experience Mod Auto Calc
     var calc_experience_mod = function() {
-        getAttrs(["cha", "experience_mod_other", "experience_mod_temp"], function(v) {
-            var base_value = Math.ceil(parseInt(v.cha)/6)-2;
+        getAttrs(["cha", "int", "experience_mod_calc", "experience_mod_other", "experience_mod_temp"], function(v) {
+            if(v.experience_mod_calc == "1") {
+                var base_value = Math.ceil(parseInt(v.int)/6)-2;
+            } else {
+                var base_value = Math.ceil(parseInt(v.cha)/6)-2;                
+            }
             
             setAttrs({
                 experience_mod_base: base_value,
@@ -230,7 +234,7 @@
             });
         });
     };
-    on("change:cha change:experience_mod_temp change:experience_mod_other", function() { calc_experience_mod(); });
+    on("change:cha change:int change:experience_mod_calc change:experience_mod_temp change:experience_mod_other", function() { calc_experience_mod(); });
 
     // Healing Rate Auto Calc
     var calc_healing_rate = function() { 
@@ -837,6 +841,10 @@
                 var base_value = parseInt(v[skill_base]);
             } else if (skill_name == "status") {
                 var base_value = parseInt(v[skill_base]);
+            } else if (skill_name == "strangeness") {
+                var base_value = parseInt(v[skill_base]);
+            } else if (skill_name == "the_soot") {
+                var base_value = parseInt(v[skill_base]);
             } else {
                 var base_value = char1_value + char2_value;
             }
@@ -1016,8 +1024,10 @@
     on("change:cha change:pow change:sing_experience change:sing_other change:sing_temp change:sing_penalty change:herculean_mod ", function() { calc_skill("sing", "@{cha}", "@{pow}", 1); });
     on("change:status_base change:status_experience change:status_temp change:status_other change:status_penalty change:herculean_mod", function() { calc_skill("status", "0", "0", 1); });
     on("change:int change:dex change:stealth_experience change:stealth_other change:stealth_temp change:stealth_penalty change:herculean_mod ", function() { calc_skill("stealth", "@{int}", "@{dex}", 1); });
+    on("change:strangeness_base change:strangeness_experience change:strangeness_temp change:strangeness_other change:strangeness_penalty change:herculean_mod", function() { calc_skill("strangeness", "0", "0", 1); });
     on("change:int change:pow change:superstition_experience change:superstition_other change:superstition_temp change:superstition_penalty change:herculean_mod ", function() { calc_skill("superstition", "@{int}", "@{pow}", 1); });
     on("change:str change:con change:swim_experience change:swim_other change:swim_temp change:swim_penalty change:herculean_mod ", function() { calc_skill("swim", "@{str}", "@{con}", 1); });
+    on("change:the_soot_base change:the_soot_experience change:the_soot_temp change:the_soot_other change:the_soot_penalty change:herculean_mod", function() { calc_skill("the_soot", "0", "0", 1); });
     on("change:pow change:theology_experience change:theology_other change:theology_temp change:theology_penalty change:theology_learned change:herculean_mod ", function() {
         getAttrs(["theology_learned"], function(v) {
             calc_skill("theology", "@{pow}", "@{pow}", v.theology_learned);
@@ -1046,6 +1056,16 @@
             if(idarray.length > 0) {
                 _.each(idarray, function(currentID, i) {
                     calc_skill("repeating_psionicpower_" + currentID, "@{pow}", "@{pow}", 1);
+                });
+            }
+        });
+    });
+    // Repeating Odd Soot Magic Spell Auto Calc
+    on("change:pow change:herculean_mod change:repeating_magicspell", function() {
+        getSectionIDs("repeating_magicspell", function(idarray) {
+            if(idarray.length > 0) {
+                _.each(idarray, function(currentID, i) {
+                    calc_skill("repeating_magicspell_" + currentID, "@{pow}", "@{pow}", 1);
                 });
             }
         });
@@ -1139,7 +1159,7 @@
         });
     });
 
-    // Passions and Dependencies Auto Calc
+    // Passions and Dependencies  and Peculiarities Auto Calc
     var calc_passion = function(skill_name) {
         var skill_total = skill_name + "_total";
         var skill_score = skill_name + "_score";
@@ -1172,7 +1192,7 @@
             setAttrs(setObj);
         });
     };
-    on("change:repeating_passion change:repeating_dependency", function(event_info) { calc_passion(event_info.triggerName); });
+    on("change:repeating_passion change:repeating_peculiarity change:repeating_dependency", function(event_info) { calc_passion(event_info.triggerName); });
 
     // Spirits Max Auto Calc
     var calc_max_spirits = function() {
@@ -1280,8 +1300,10 @@
         calc_skill("sing", "@{cha}", "@{pow}", 1);
         calc_skill("status", "0", "0", 1);
         calc_skill("stealth", "@{int}", "@{dex}", 1);
+        calc_skill("strangeness", "0", "0", 1);
         calc_skill("superstition", "@{int}", "@{pow}", 1);
         calc_skill("swim", "@{str}", "@{con}", 1);
+        calc_skill("the_soot", "0", "0", 1);
         getAttrs(["theology_learned"], function(v) {
             calc_skill("theology", "@{pow}", "@{pow}", v.theology_learned);
         });
@@ -1310,6 +1332,13 @@
             if(idarray.length > 0) {
                 _.each(idarray, function(currentID, i) {
                     calc_skill("repeating_psionicpower_" + currentID, "@{pow}", "@{pow}", 1);
+                });
+            }
+        });
+        getSectionIDs("repeating_magicspell", function(idarray) {
+            if(idarray.length > 0) {
+                _.each(idarray, function(currentID, i) {
+                    calc_skill("repeating_magicspell_" + currentID, "@{pow}", "@{pow}", 1);
                 });
             }
         });
@@ -1391,6 +1420,13 @@
             if(idarray.length > 0) {
                 _.each(idarray, function(currentID, i) {
                     calc_passion("repeating_dependency_" + currentID);
+                });
+            }
+        });
+        getSectionIDs("repeating_peculiarity", function(idarray) {
+            if(idarray.length > 0) {
+                _.each(idarray, function(currentID, i) {
+                    calc_passion("repeating_peculiarity_" + currentID);
                 });
             }
         });
@@ -2374,6 +2410,11 @@
                 setAttrs({version: "2.3"});
                 versioning();
             }
+            else if(v["version"] === "2.3") {
+                console.log("upgrading to v2.4");
+                setAttrs({version: "2.4"});
+                versioning();
+            }
             else {
                 console.log("Sheet fully updated");
             }
@@ -2381,7 +2422,7 @@
     }
     
     var campaign_options = function() {
-        getAttrs(["setting_option", "luck_points_rank_option", "herculean_mod_option", "battle_units_enabled_option", "vehicle_type_option", "extended_conflict_enabled_option", "simplified_combat_enabled_option", "action_points_calc_option", "magic_points_enabled_option", "power_points_enabled_option", "prana_points_enabled_option", "tenacity_enabled_option", "spirits_enabled_option", "luther_arkwright_style_option", "m_space_style_option", "boating_standard_option", "status_standard_option", "superstition_standard_option", "linguistics_enabled_option", "dependencies_enabled_option", "firearms_enabled_option", "reach_enabled_option", "affiliations_enabled_option", "ms_psionics_enabled_option", "roman_magic_enabled_option", "arcane_magic_enabled_option", "divine_magic_enabled_option", "folk_magic_enabled_option", "superpowers_enabled_option", "fae_powers_enabled_option", "folk_magic_range_multiplier_option", "animism_enabled_option", "mysticism_enabled_option", "mythras_psionics_enabled_option", "sorcery_enabled_option", "theism_enabled_option", "max_devotional_pool_based_on_option"], function(v) {
+        getAttrs(["setting_option", "luck_points_rank_option", "herculean_mod_option", "battle_units_enabled_option", "vehicle_type_option", "extended_conflict_enabled_option", "simplified_combat_enabled_option", "action_points_calc_option", "magic_points_enabled_option", "power_points_enabled_option", "prana_points_enabled_option", "tenacity_enabled_option", "spirits_enabled_option", "luther_arkwright_style_option", "m_space_style_option", "odd_soot_style_option", "boating_standard_option", "status_standard_option", "strangeness_standard_option", "superstition_standard_option", "the_soot_standard_option", "linguistics_enabled_option", "dependencies_enabled_option", "peculiarities_enabled_option", "firearms_enabled_option", "reach_enabled_option", "affiliations_enabled_option", "ms_psionics_enabled_option", "os_magic_enabled_option", "roman_magic_enabled_option", "arcane_magic_enabled_option", "divine_magic_enabled_option", "folk_magic_enabled_option", "superpowers_enabled_option", "fae_powers_enabled_option", "folk_magic_range_multiplier_option", "animism_enabled_option", "mysticism_enabled_option", "mythras_psionics_enabled_option", "sorcery_enabled_option", "theism_enabled_option", "max_devotional_pool_based_on_option"], function(v) {
             var newoptions = {};
             // Default Setting Configs
             var setting_configs = {
@@ -2399,12 +2440,16 @@
                 tenacity_enabled: "0",
                 boating_standard: "1",
                 status_standard: "0",
+                strangeness_standard: "0",
                 superstition_standard: "0",
+                the_soot_standard: "0",
                 linguistics_enabled: "0",
                 dependencies_enabled: "0",
+                peculiarities_enabled: "0",
                 firearms_enabled: "0",
                 reach_enabled: "1",
                 ms_psionics_enabled: "0",
+                os_magic_enabled: "0",
                 roman_magic_enabled: "0",
                 arcane_magic_enabled: "0",
                 divine_magic_enabled: "0",
@@ -2484,6 +2529,26 @@
                 setting_configs["firearms_enabled"] = "1";
                 setting_configs["reach_enabled"] = "0";
                 setting_configs["ms_psionics_enabled"] = "1";
+                setting_configs["folk_magic_enabled"] = "0";
+                setting_configs["animism_enabled"] = "0";
+                setting_configs["mysticism_enabled"] = "0";
+                setting_configs["sorcery_enabled"] = "0";
+                setting_configs["theism_enabled"] = "0";
+                setting_configs["vehicle_type"] = "mspace";
+            } else if(v["setting_option"] === "odd_soot") {
+                setting_configs["herculean_mod"] = ".2";
+                setting_configs["spirits_enabled"] = "0";
+                setting_configs["magic_points_enabled"] = "0";
+                setting_configs["power_points_enabled"] = "1";
+                setting_configs["sheet_style"] = "odd_soot";
+                setting_configs["extended_conflict_enabled"] = "1",
+                setting_configs["action_points_calc"] = "set_2";
+                setting_configs["firearms_enabled"] = "1";
+                setting_configs["reach_enabled"] = "0";
+                setting_configs["os_magic_enabled"] = "1";
+                setting_configs["peculiarities_enabled"] = "1";
+                setting_configs["the_soot_standard"] = "1";
+                setting_configs["strangeness_standard"] = "1";
                 setting_configs["folk_magic_enabled"] = "0";
                 setting_configs["animism_enabled"] = "0";
                 setting_configs["mysticism_enabled"] = "0";
@@ -2630,12 +2695,26 @@
             } else {
                 newoptions["status_standard"] = v["status_standard_option"];
             }
+
+            // Strangeness Standard
+            if(v["strangeness_standard_option"] === "default") {
+                newoptions["strangeness_standard"] = setting_configs["strangeness_standard"];
+            } else {
+                newoptions["strangeness_standard"] = v["strangeness_standard_option"];
+            }
             
             // Superstition Standard
             if(v["superstition_standard_option"] === "default") {
                 newoptions["superstition_standard"] = setting_configs["superstition_standard"];
             } else {
                 newoptions["superstition_standard"] = v["superstition_standard_option"];
+            }
+
+            // The Soot Standard
+            if(v["the_soot_standard_option"] === "default") {
+                newoptions["the_soot_standard"] = setting_configs["the_soot_standard"];
+            } else {
+                newoptions["the_soot_standard"] = v["the_soot_standard_option"];
             }
             
             // Linguistics Enabled
@@ -2645,11 +2724,18 @@
                 newoptions["linguistics_enabled"] = v["linguistics_enabled_option"];
             }
             
-            // Dependencie Enabled
+            // Dependencies Enabled
             if(v["dependencies_enabled_option"] === "default") {
                 newoptions["dependencies_enabled"] = setting_configs["dependencies_enabled"];
             } else {
                 newoptions["dependencies_enabled"] = v["dependencies_enabled_option"];
+            }
+
+            // Peculiarities Enabled
+            if(v["peculiarities_enabled_option"] === "default") {
+                newoptions["peculiarities_enabled"] = setting_configs["peculiarities_enabled"];
+            } else {
+                newoptions["peculiarities_enabled"] = v["peculiarities_enabled_option"];
             }
             
             // Firearms Enabled
@@ -2671,6 +2757,13 @@
                 newoptions["ms_psionics_enabled"] = setting_configs["ms_psionics_enabled"];
             } else {
                 newoptions["ms_psionics_enabled"] = v["ms_psionics_enabled_option"];
+            }
+
+            // Odd Soot Magic Enabled
+            if(v["os_magic_enabled_option"] === "default") {
+                newoptions["os_magic_enabled"] = setting_configs["os_magic_enabled"];
+            } else {
+                newoptions["os_magic_enabled"] = v["os_magic_enabled_option"];
             }
             
             // Roman Magic Enabled
@@ -4688,6 +4781,7 @@
                     movement_rate_species_swim: characterData["attributes"]["movement"]
                 });
             }
+            /*
             if (typeof characterData["attributes"]["action_points"] !== 'undefined') {
                 setAttrs({
                     action_points: characterData["attributes"]["action_points"],
@@ -4763,6 +4857,7 @@
                     tenacity_mook10: characterData["attributes"]["tenacity"]
                 });
             }
+            */
             if (typeof characterData["attributes"]["strike_rank"] !== 'undefined') {
                 if (characterData["attributes"]["strike_rank"].includes("-")) {
                     var strike_rank_other = characterData["attributes"]["strike_rank"].split("-")[1].replace(')','');
@@ -4800,7 +4895,7 @@
                 if (typeof characterData["hit_locations"][0]["ap"] !== 'undefined') {
                     setAttrs({location1_armor_ap: characterData["hit_locations"][0]["ap"]});
                 }
-                var simplified_hp_val = Math.ceil((con+siz)/2);
+                /*var simplified_hp_val = Math.ceil((con+siz)/2);
                 setAttrs({
                     simplified_hp: simplified_hp_val,
                     simplified_hp_mook1: simplified_hp_val,
@@ -4813,7 +4908,7 @@
                     simplified_hp_mook8: simplified_hp_val,
                     simplified_hp_mook9: simplified_hp_val,
                     simplified_hp_mook10: simplified_hp_val
-                });
+                });*/
             } else {
                 var base_hp = Math.ceil((con+siz)/5);
                 
@@ -4825,6 +4920,7 @@
                     } else {
                         table[1] = parseInt(table[1]);
                     }
+                    /*
                     setAttrs({
                         location1_table_start: table[0],
                         location1_table_end: table[1],
@@ -4844,7 +4940,17 @@
                         location1_hp_mook9: characterData["hit_locations"][0]["hp"],
                         location1_hp_mook10: characterData["hit_locations"][0]["hp"]
                     });
+                    */
+                    setAttrs({
+                        location1_table_start: table[0],
+                        location1_table_end: table[1],
+                        location1_name: characterData["hit_locations"][0]["name"],
+                        location1_armor_ap: characterData["hit_locations"][0]["ap"],
+                        location1_armor_ap_max: characterData["hit_locations"][0]["ap"],
+                        location1_hp_max_base_mod: characterData["hit_locations"][0]["hp"] - base_hp
+                    });
                 } else {
+                    /*
                     setAttrs({
                         location1_table_start: "0",
                         location1_table_end: "0",
@@ -4864,6 +4970,15 @@
                         location1_hp_mook9: "0",
                         location1_hp_mook10: "0"
                     });
+                    */
+                    setAttrs({
+                        location1_table_start: "0",
+                        location1_table_end: "0",
+                        location1_name: " ",
+                        location1_armor_ap: "0",
+                        location1_armor_ap_max: "0",
+                        location1_hp_max_base_mod: 0 - base_hp
+                    });
                 }
                 
                 if (typeof characterData["hit_locations"][1] !== 'undefined') {
@@ -4874,6 +4989,7 @@
                     } else {
                         table[1] = parseInt(table[1]);
                     }
+                    /*
                     setAttrs({
                         location2_table_start: table[0],
                         location2_table_end: table[1],
@@ -4893,7 +5009,17 @@
                         location2_hp_mook9: characterData["hit_locations"][1]["hp"],
                         location2_hp_mook10: characterData["hit_locations"][1]["hp"]
                     });
+                    */
+                    setAttrs({
+                        location2_table_start: table[0],
+                        location2_table_end: table[1],
+                        location2_name: characterData["hit_locations"][1]["name"],
+                        location2_armor_ap: characterData["hit_locations"][1]["ap"],
+                        location2_armor_ap_max: characterData["hit_locations"][1]["ap"],
+                        location2_hp_max_base_mod: characterData["hit_locations"][1]["hp"] - base_hp,
+                    });
                 } else {
+                    /*
                     setAttrs({
                         location2_table_start: "0",
                         location2_table_end: "0",
@@ -4913,6 +5039,15 @@
                         location2_hp_mook9: "0",
                         location2_hp_mook10: "0"
                     });
+                    */
+                    setAttrs({
+                        location2_table_start: "0",
+                        location2_table_end: "0",
+                        location2_name: " ",
+                        location2_armor_ap: "0",
+                        location2_armor_ap_max: "0",
+                        location2_hp_max_base_mod: 0 - base_hp
+                    });
                 }
                 
                 if (typeof characterData["hit_locations"][2] !== 'undefined') {
@@ -4923,6 +5058,7 @@
                     } else {
                         table[1] = parseInt(table[1]);
                     }
+                    /*
                     setAttrs({
                         location3_table_start: table[0],
                         location3_table_end: table[1],
@@ -4942,7 +5078,17 @@
                         location3_hp_mook9: characterData["hit_locations"][2]["hp"],
                         location3_hp_mook10: characterData["hit_locations"][2]["hp"]
                     });
+                    */
+                    setAttrs({
+                        location3_table_start: table[0],
+                        location3_table_end: table[1],
+                        location3_name: characterData["hit_locations"][2]["name"],
+                        location3_armor_ap: characterData["hit_locations"][2]["ap"],
+                        location3_armor_ap_max: characterData["hit_locations"][2]["ap"],
+                        location3_hp_max_base_mod: characterData["hit_locations"][2]["hp"] - base_hp
+                    });
                 } else {
+                    /*
                     setAttrs({
                         location3_table_start: "0",
                         location3_table_end: "0",
@@ -4961,6 +5107,15 @@
                         location3_hp_mook9: "0",
                         location3_hp_mook10: "0"
                     });
+                    */
+                    setAttrs({
+                        location3_table_start: "0",
+                        location3_table_end: "0",
+                        location3_name: " ",
+                        location3_armor_ap: "0",
+                        location3_armor_ap_max: "0",
+                        location3_hp_max_base_mod: 0 - base_hp
+                    });
                 }
                 
                 if (typeof characterData["hit_locations"][3] !== 'undefined') {
@@ -4971,6 +5126,7 @@
                     } else {
                         table[1] = parseInt(table[1]);
                     }
+                    /*
                     setAttrs({
                         location4_table_start: table[0],
                         location4_table_end: table[1],
@@ -4990,7 +5146,17 @@
                         location4_hp_mook9: characterData["hit_locations"][3]["hp"],
                         location4_hp_mook10: characterData["hit_locations"][3]["hp"]
                     });
+                    */
+                    setAttrs({
+                        location4_table_start: table[0],
+                        location4_table_end: table[1],
+                        location4_name: characterData["hit_locations"][3]["name"],
+                        location4_armor_ap: characterData["hit_locations"][3]["ap"],
+                        location4_armor_ap_max: characterData["hit_locations"][3]["ap"],
+                        location4_hp_max_base_mod: characterData["hit_locations"][3]["hp"] - base_hp
+                    });
                 } else {
+                    /*
                     setAttrs({
                         location4_table_start: "0",
                         location4_table_end: "0",
@@ -5009,6 +5175,15 @@
                         location4_hp_mook9: "0",
                         location4_hp_mook10: "0"
                     });
+                    */
+                    setAttrs({
+                        location4_table_start: "0",
+                        location4_table_end: "0",
+                        location4_name: " ",
+                        location4_armor_ap: "0",
+                        location4_armor_ap_max: "0",
+                        location4_hp_max_base_mod: 0 - base_hp
+                    });
                 }
                 
                 if (typeof characterData["hit_locations"][4] !== 'undefined') {
@@ -5019,6 +5194,7 @@
                     } else {
                         table[1] = parseInt(table[1]);
                     }
+                    /*
                     setAttrs({
                         location5_table_start: table[0],
                         location5_table_end: table[1],
@@ -5038,7 +5214,17 @@
                         location5_hp_mook9: characterData["hit_locations"][4]["hp"],
                         location5_hp_mook10: characterData["hit_locations"][4]["hp"]
                     });
+                    */
+                    setAttrs({
+                        location5_table_start: table[0],
+                        location5_table_end: table[1],
+                        location5_name: characterData["hit_locations"][4]["name"],
+                        location5_armor_ap: characterData["hit_locations"][4]["ap"],
+                        location5_armor_ap_max: characterData["hit_locations"][4]["ap"],
+                        location5_hp_max_base_mod: characterData["hit_locations"][4]["hp"] - base_hp
+                    });
                 } else {
+                    /*
                     setAttrs({
                         location5_table_start: "0",
                         location5_table_end: "0",
@@ -5057,6 +5243,15 @@
                         location5_hp_mook9: "0",
                         location5_hp_mook10: "0"
                     });
+                    */
+                    setAttrs({
+                        location5_table_start: "0",
+                        location5_table_end: "0",
+                        location5_name: " ",
+                        location5_armor_ap: "0",
+                        location5_armor_ap_max: "0",
+                        location5_hp_max_base_mod: 0 - base_hp
+                    });
                 }
                 
                 if (typeof characterData["hit_locations"][5] !== 'undefined') {
@@ -5067,6 +5262,7 @@
                     } else {
                         table[1] = parseInt(table[1]);
                     }
+                    /*
                     setAttrs({
                         location6_table_start: table[0],
                         location6_table_end: table[1],
@@ -5086,7 +5282,17 @@
                         location6_hp_mook9: characterData["hit_locations"][5]["hp"],
                         location6_hp_mook10: characterData["hit_locations"][5]["hp"]
                     });
+                    */
+                    setAttrs({
+                        location6_table_start: table[0],
+                        location6_table_end: table[1],
+                        location6_name: characterData["hit_locations"][5]["name"],
+                        location6_armor_ap: characterData["hit_locations"][5]["ap"],
+                        location6_armor_ap_max: characterData["hit_locations"][5]["ap"],
+                        location6_hp_max_base_mod: characterData["hit_locations"][5]["hp"] - base_hp
+                    });
                 } else {
+                    /*
                     setAttrs({
                         location6_table_start: "0",
                         location6_table_end: "0",
@@ -5105,6 +5311,15 @@
                         location6_hp_mook9: "0",
                         location6_hp_mook10: "0"
                     });
+                    */
+                    setAttrs({
+                        location6_table_start: "0",
+                        location6_table_end: "0",
+                        location6_name: " ",
+                        location6_armor_ap: "0",
+                        location6_armor_ap_max: "0",
+                        location6_hp_max_base_mod: 0 - base_hp
+                    });
                 }
                 
                 if (typeof characterData["hit_locations"][6] !== 'undefined') {
@@ -5115,6 +5330,7 @@
                     } else {
                         table[1] = parseInt(table[1]);
                     }
+                    /*
                     setAttrs({
                         location7_table_start: table[0],
                         location7_table_end: table[1],
@@ -5134,7 +5350,17 @@
                         location7_hp_mook9: characterData["hit_locations"][6]["hp"],
                         location7_hp_mook10: characterData["hit_locations"][6]["hp"]
                     });
+                    */
+                    setAttrs({
+                        location7_table_start: table[0],
+                        location7_table_end: table[1],
+                        location7_name: characterData["hit_locations"][6]["name"],
+                        location7_armor_ap: characterData["hit_locations"][6]["ap"],
+                        location7_armor_ap_max: characterData["hit_locations"][6]["ap"],
+                        location7_hp_max_base_mod: characterData["hit_locations"][6]["hp"] - base_hp
+                    });
                 } else {
+                    /*
                     setAttrs({
                         location7_table_start: "0",
                         location7_table_end: "0",
@@ -5153,6 +5379,15 @@
                         location7_hp_mook9: "0",
                         location7_hp_mook10: "0"
                     });
+                    */
+                    setAttrs({
+                        location7_table_start: "0",
+                        location7_table_end: "0",
+                        location7_name: " ",
+                        location7_armor_ap: "0",
+                        location7_armor_ap_max: "0",
+                        location7_hp_max_base_mod: 0 - base_hp
+                    });
                 }
                 
                 if (typeof characterData["hit_locations"][7] !== 'undefined') {
@@ -5163,6 +5398,7 @@
                     } else {
                         table[1] = parseInt(table[1]);
                     }
+                    /*
                     setAttrs({
                         location8_table_start: table[0],
                         location8_table_end: table[1],
@@ -5182,7 +5418,17 @@
                         location8_hp_mook9: characterData["hit_locations"][7]["hp"],
                         location8_hp_mook10: characterData["hit_locations"][7]["hp"]
                     });
+                    */
+                    setAttrs({
+                        location8_table_start: table[0],
+                        location8_table_end: table[1],
+                        location8_name: characterData["hit_locations"][7]["name"],
+                        location8_armor_ap: characterData["hit_locations"][7]["ap"],
+                        location8_armor_ap_max: characterData["hit_locations"][7]["ap"],
+                        location8_hp_max_base_mod: characterData["hit_locations"][7]["hp"] - base_hp
+                    });
                 } else {
+                    /*
                     setAttrs({
                         location8_table_start: "0",
                         location8_table_end: "0",
@@ -5201,6 +5447,15 @@
                         location8_hp_mook9: "0",
                         location8_hp_mook10: "0"
                     });
+                    */
+                    setAttrs({
+                        location8_table_start: "0",
+                        location8_table_end: "0",
+                        location8_name: " ",
+                        location8_armor_ap: "0",
+                        location8_armor_ap_max: "0",
+                        location8_hp_max_base_mod: 0 - base_hp
+                    });
                 }
                 
                 if (typeof characterData["hit_locations"][8] !== 'undefined') {
@@ -5211,6 +5466,7 @@
                     } else {
                         table[1] = parseInt(table[1]);
                     }
+                    /*
                     setAttrs({
                         location9_table_start: table[0],
                         location9_table_end: table[1],
@@ -5230,7 +5486,17 @@
                         location9_hp_mook9: characterData["hit_locations"][8]["hp"],
                         location9_hp_mook10: characterData["hit_locations"][8]["hp"]
                     });
+                    */
+                    setAttrs({
+                        location9_table_start: table[0],
+                        location9_table_end: table[1],
+                        location9_name: characterData["hit_locations"][8]["name"],
+                        location9_armor_ap: characterData["hit_locations"][8]["ap"],
+                        location9_armor_ap_max: characterData["hit_locations"][8]["ap"],
+                        location9_hp_max_base_mod: characterData["hit_locations"][8]["hp"] - base_hp
+                    });
                 } else {
+                    /*
                     setAttrs({
                         location9_table_start: "0",
                         location9_table_end: "0",
@@ -5249,6 +5515,15 @@
                         location9_hp_mook9: "0",
                         location9_hp_mook10: "0"
                     });
+                    */
+                    setAttrs({
+                        location9_table_start: "0",
+                        location9_table_end: "0",
+                        location9_name: " ",
+                        location9_armor_ap: "0",
+                        location9_armor_ap_max: "0",
+                        location9_hp_max_base_mod: 0 - base_hp
+                    });
                 }
                 
                 if (typeof characterData["hit_locations"][9] !== 'undefined') {
@@ -5259,6 +5534,7 @@
                     } else {
                         table[1] = parseInt(table[1]);
                     }
+                    /*
                     setAttrs({
                         location10_table_start: table[0],
                         location10_table_end: table[1],
@@ -5278,7 +5554,17 @@
                         location10_hp_mook9: characterData["hit_locations"][9]["hp"],
                         location10_hp_mook10: characterData["hit_locations"][9]["hp"]
                     });
+                    */
+                    setAttrs({
+                        location10_table_start: table[0],
+                        location10_table_end: table[1],
+                        location10_name: characterData["hit_locations"][9]["name"],
+                        location10_armor_ap: characterData["hit_locations"][9]["ap"],
+                        location10_armor_ap_max: characterData["hit_locations"][9]["ap"],
+                        location10_hp_max_base_mod: characterData["hit_locations"][9]["hp"] - base_hp
+                    });
                 } else {
+                    /*
                     setAttrs({
                         location10_table_start: "0",
                         location10_table_end: "0",
@@ -5297,6 +5583,15 @@
                         location10_hp_mook9: "0",
                         location10_hp_mook10: "0"
                     });
+                    */
+                    setAttrs({
+                        location10_table_start: "0",
+                        location10_table_end: "0",
+                        location10_name: " ",
+                        location10_armor_ap: "0",
+                        location10_armor_ap_max: "0",
+                        location10_hp_max_base_mod: 0 - base_hp
+                    });
                 }
                 
                 if (typeof characterData["hit_locations"][10] !== 'undefined') {
@@ -5307,6 +5602,7 @@
                     } else {
                         table[1] = parseInt(table[1]);
                     }
+                    /*
                     setAttrs({
                         location11_table_start: table[0],
                         location11_table_end: table[1],
@@ -5326,7 +5622,17 @@
                         location11_hp_mook9: characterData["hit_locations"][10]["hp"],
                         location11_hp_mook10: characterData["hit_locations"][10]["hp"]
                     });
+                    */
+                    setAttrs({
+                        location11_table_start: table[0],
+                        location11_table_end: table[1],
+                        location11_name: characterData["hit_locations"][10]["name"],
+                        location11_armor_ap: characterData["hit_locations"][10]["ap"],
+                        location11_armor_ap_max: characterData["hit_locations"][10]["ap"],
+                        location11_hp_max_base_mod: characterData["hit_locations"][10]["hp"] - base_hp
+                    });
                 } else {
+                    /*
                     setAttrs({
                         location11_table_start: "0",
                         location11_table_end: "0",
@@ -5345,6 +5651,15 @@
                         location11_hp_mook9: "0",
                         location11_hp_mook10: "0"
                     });
+                    */
+                    setAttrs({
+                        location11_table_start: "0",
+                        location11_table_end: "0",
+                        location11_name: " ",
+                        location11_armor_ap: "0",
+                        location11_armor_ap_max: "0",
+                        location11_hp_max_base_mod: 0 - base_hp
+                    });
                 }
                 
                 if (typeof characterData["hit_locations"][11] !== 'undefined') {
@@ -5355,6 +5670,7 @@
                     } else {
                         table[1] = parseInt(table[1]);
                     }
+                    /*
                     setAttrs({
                         location12_table_start: table[0],
                         location12_table_end: table[1],
@@ -5374,7 +5690,17 @@
                         location12_hp_mook9: characterData["hit_locations"][11]["hp"],
                         location12_hp_mook10: characterData["hit_locations"][11]["hp"]
                     });
+                    */
+                    setAttrs({
+                        location12_table_start: table[0],
+                        location12_table_end: table[1],
+                        location12_name: characterData["hit_locations"][11]["name"],
+                        location12_armor_ap: characterData["hit_locations"][11]["ap"],
+                        location12_armor_ap_max: characterData["hit_locations"][11]["ap"],
+                        location12_hp_max_base_mod: characterData["hit_locations"][11]["hp"] - base_hp
+                    });
                 } else {
+                    /*
                     setAttrs({
                         location12_table_start: "0",
                         location12_table_end: "0",
@@ -5392,6 +5718,15 @@
                         location12_hp_mook8: "0",
                         location12_hp_mook9: "0",
                         location12_hp_mook10: "0"
+                    });
+                    */
+                    setAttrs({
+                        location12_table_start: "0",
+                        location12_table_end: "0",
+                        location12_name: " ",
+                        location12_armor_ap: "0",
+                        location12_armor_ap_max: "0",
+                        location12_hp_max_base_mod: 0 - base_hp
                     });
                 }
             }
@@ -5624,10 +5959,14 @@
                         skillattrs["status_experience"] = skillValue;
                     } else if (skillKey.toLowerCase() == "stealth") {
                         skillattrs["stealth_experience"] = skillValue - dex - int;
+                    } else if (skillKey.toLowerCase() == "strangeness") {
+                        skillattrs["strangeness_experience"] = skillValue;
                     } else if (skillKey.toLowerCase() == "superstition") {
                         skillattrs["superstition_experience"] = skillValue - ((21 - int) + pow);
                     } else if (skillKey.toLowerCase() == "swim") {
                         skillattrs["swim_experience"] = skillValue - str - con;
+                    } else if (skillKey.toLowerCase() == "the soot") {
+                        skillattrs["the_soot_experience"] = skillValue;
                     } else if (skillKey.toLowerCase() == "unarmed") {
                         skillattrs["unarmed_experience"] = skillValue - str - dex;
                     } else if (skillKey.toLowerCase() == "willpower") {
@@ -5914,6 +6253,10 @@
                         skillattrs["repeating_dependency_" + skillid + "_name"] = skillKey;
                         skillattrs["repeating_dependency_" + skillid + "_score"] = skillValue;
                         skillattrs["repeating_dependency_" + skillid + "_details"] = 0;
+                    } else if (skillKey.toLowerCase().indexOf("peculiarity") !== -1) {
+                        skillattrs["repeating_peculiarity_" + skillid + "_name"] = skillKey;
+                        skillattrs["repeating_peculiarity_" + skillid + "_score"] = skillValue;
+                        skillattrs["repeating_peculiarity_" + skillid + "_details"] = 0;
                     } else if (skillKey.toLowerCase().indexOf("affiliation") !== -1) {
                         skillattrs["repeating_affiliation_" + skillid + "_name"] = skillKey;
                         skillattrs["repeating_affiliation_" + skillid + "_experience"] = skillValue - 40;
@@ -6126,7 +6469,7 @@
     });
     
     //Set campaign options if any change
-    on("change:setting_option change:vehicle_type_option change:battle_units_enabled_option change:extended_conflict_enabled_option change:simplified_combat_enabled_option change:luck_points_rank_option change:herculean_mod_option change:action_points_calc_option change:magic_points_enabled_option change:power_points_enabled_option change:prana_points_enabled_option change:spirits_enabled_option change:tenacity_enabled_option change:status_standard_option change:superstition_standard_option change:boating_standard_option change:linguistics_enabled_option change:dependencies_enabled_option change:firearms_enabled_option change:reach_enabled_option change:affiliations_enabled_option change:roman_magic_enabled_option change:arcane_magic_enabled_option change:divine_magic_enabled_option change:superpowers_enabled_option change:fae_powers_enabled_option change:folk_magic_enabled_option change:folk_magic_range_multiplier_option change:animism_enabled_option change:mysticism_enabled_option change:mythras_psionics_enabled_option change:ms_psionics_enabled_option change:sorcery_enabled_option change:theism_enabled_option change:max_devotional_pool_based_on_option", function() {
+    on("change:setting_option change:vehicle_type_option change:battle_units_enabled_option change:extended_conflict_enabled_option change:simplified_combat_enabled_option change:luck_points_rank_option change:herculean_mod_option change:action_points_calc_option change:magic_points_enabled_option change:power_points_enabled_option change:prana_points_enabled_option change:spirits_enabled_option change:tenacity_enabled_option change:status_standard_option change:strangeness_standard_option change:superstition_standard_option change:the_soot_standard_option change:boating_standard_option change:linguistics_enabled_option change:dependencies_enabled_option change:peculiarities_enabled_option change:firearms_enabled_option change:reach_enabled_option change:affiliations_enabled_option change:roman_magic_enabled_option change:arcane_magic_enabled_option change:divine_magic_enabled_option change:superpowers_enabled_option change:fae_powers_enabled_option change:folk_magic_enabled_option change:folk_magic_range_multiplier_option change:animism_enabled_option change:mysticism_enabled_option change:mythras_psionics_enabled_option change:ms_psionics_enabled_option change:os_magic_enabled_option change:sorcery_enabled_option change:theism_enabled_option change:max_devotional_pool_based_on_option", function() {
         console.log("Setting campaign options")
         campaign_options();
     });